+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/transforms/Tanh.java
@@ -27,7 +27,7 @@ public class Tanh extends BaseElementWiseOp {
         }
         else  {
             float d = (float) value;
-            return  Math.tanh(d);
+            return  (float)  Math.tanh(d);
         }
     }
 }
